+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2021-06-26
 * Fixed a bug that caused the `/api/v1/list` API endpoint to return HTTP 500 errors in production environments.
+* Fixed a bug that caused "ghost" sessions/worlds to remain visible in the web interface after they were closed.
 
 ## 2021-06-25
 * The following API endpoints have been added, accessible at `/api/v1/<client_id>`: